+++ examples/graphiql-parcel/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.10-alpha.1](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.10-alpha.0...graphiql-parcel-example@1.1.10-alpha.1) (2020-08-12)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
+## [1.1.10-alpha.0](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.9...graphiql-parcel-example@1.1.10-alpha.0) (2020-08-10)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
+## [1.1.9](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.5...graphiql-parcel-example@1.1.9) (2020-08-06)
+
+**Note:** Version bump only for package graphiql-parcel-example
+
 ## [1.1.5](https://github.com/graphql/graphiql/compare/graphiql-parcel-example@1.1.4...graphiql-parcel-example@1.1.5) (2020-06-11)
 
 **Note:** Version bump only for package graphiql-parcel-example